feat(web): move leaderboard to /leaderboard and add Users tabs
Moves leaderboard from /agents/leaderboard to /leaderboard as a
top-level route — it covers both agents and components.

Adds a Users sub-tab to both the Agents and Components sections
that shows per-user aggregate download totals and item counts.
Components leaderboard now shows individual components ranked by
downloads instead of the old per-user MCP count view.

Time window selector now applies to both Agents and Components.
